ARGO-3145 Reports internal POEM API methods tighten

+++ poem/Poem/api/internal_views/reports.py
@@ -3,6 +3,7 @@ from django.contrib.contenttypes.models import ContentType
 
 from Poem.api import serializers
 from Poem.api.internal_views.utils import sync_webapi
+from Poem.api.internal_views.users import get_all_groups, get_groups_for_user
 from Poem.api.views import NotFound
 from Poem.helpers.history_helpers import create_profile_history
 from Poem.poem import models as poem_models
@@ -11,12 +12,36 @@ from rest_framework import status
 from rest_framework.authentication import SessionAuthentication
 from rest_framework.response import Response
 from rest_framework.views import APIView
+from Poem.users.models import CustUser
+from .utils import error_response
+
+
+def user_groups(username):
+    user = CustUser.objects.get(username=username)
+    groups = get_groups_for_user(user)
+    return groups
 
 
 class ListReports(APIView):
     authentication_classes = (SessionAuthentication,)
 
+    def _check_onchange_groupperm(self, groupname, user):
+        groups = user_groups(user)
+        if (not user.is_superuser and
+            groupname not in groups['reports']):
+            return False
+        return True
+
     def post(self, request):
+        user = request.user
+
+        if not self._check_onchange_groupperm(request.data['groupname'], user):
+            return error_response(
+                status_code=status.HTTP_401_UNAUTHORIZED,
+                detail='You do not have permission to add '
+                        'report.'
+            )
+
         serializer = serializers.ReportsSerializer(data=request.data)
 
         if serializer.is_valid():
@@ -64,6 +89,15 @@ class ListReports(APIView):
             return Response(serializer.data)
 
     def put(self, request):
+        user = request.user
+
+        if not self._check_onchange_groupperm(request.data['groupname'], user):
+            return error_response(
+                status_code=status.HTTP_401_UNAUTHORIZED,
+                detail='You do not have permission to change'
+                        'report.'
+            )
+
         if request.data['apiid']:
             report = poem_models.Reports.objects.get(
                 apiid=request.data['apiid']
@@ -87,11 +121,16 @@ class ListReports(APIView):
             )
 
     def delete(self, request, report_name=None):
+        user = request.user
+
         if report_name:
             try:
                 report = poem_models.Reports.objects.get(
                     apiid=report_name
                 )
+
+                if report.groupname not in user_groups(user)['reports']:
+                    return Response(data=None, status=status.HTTP_401_UNAUTHORIZED)
 
                 report.delete()
